// src/entities/UserProfile/types.ts
export interface IUserProfile {
  // quantity_posts: string;
  // quantity_followers: string;
  // quantity_following: string;

  about: string;
  avatar: string;
  backgroundImage: string;
  dateJoined: string;
  email: string;
  firstName: string;
  id: string;
  lastName: string;
  nickname: string;
}